backend-agent-as-tool/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Retail Forecasting System with AI Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.5.1",
    "python-decouple>=3.8",
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "watchdog>=3.0.0",
    "prophet>=1.1.5",
    "statsmodels>=0.14.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
